feedbackform types organized
