+++ src/types/FeedbackEntry.ts
@@ -1,0 +1,6 @@
+export type FeedbackForm = {
+  name: string | null;
+  email: string | null;
+  feedback?: string | null;
+  interest?: boolean | null;
+};
